pyproject.toml
[project]
name = "nlp-business"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "aiohttp>=3.9",
    "rapidfuzz>=3.9",
    "python-dotenv>=1.2.2",
]
